Move App.dart
