+++ lib/entrypoint.dart
@@ -66,117 +66,3 @@ Future<void> entrypoint() async {
 }
 
 final navigatorKey = GlobalKey<NavigatorState>();
-
-class App extends StatelessWidget {
-  const App({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      navigatorKey: navigatorKey,
-      navigatorObservers: [FirebaseAnalyticsObserver(analytics: firebaseAnalytics)],
-      theme: ThemeData(
-        appBarTheme: const AppBarTheme(
-          systemOverlayStyle: SystemUiOverlayStyle.dark,
-          centerTitle: true,
-          color: PilllColors.white,
-          elevation: 3,
-        ),
-        textSelectionTheme: const TextSelectionThemeData(
-          cursorColor: PilllColors.secondary,
-        ),
-        primaryColor: PilllColors.secondary,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        cupertinoOverrideTheme: const NoDefaultCupertinoThemeData(
-            textTheme: CupertinoTextThemeData(
-                textStyle: TextStyle(
-          fontFamily: FontFamily.number,
-          fontWeight: FontWeight.w500,
-          fontSize: 24,
-        ))),
-        buttonTheme: const ButtonThemeData(
-          buttonColor: PilllColors.primary,
-          disabledColor: PilllColors.disable,
-          textTheme: ButtonTextTheme.primary,
-          colorScheme: ColorScheme.light(
-            primary: PilllColors.secondary,
-            secondary: PilllColors.accent,
-          ),
-        ),
-        switchTheme: SwitchThemeData(
-          thumbColor: MaterialStateProperty.resolveWith<Color?>((Set<MaterialState> states) {
-            if (states.contains(MaterialState.disabled)) {
-              return null;
-            }
-            if (states.contains(MaterialState.selected)) {
-              return PilllColors.secondary;
-            }
-            return null;
-          }),
-          trackColor: MaterialStateProperty.resolveWith<Color?>((Set<MaterialState> states) {
-            if (states.contains(MaterialState.disabled)) {
-              return null;
-            }
-            if (states.contains(MaterialState.selected)) {
-              return PilllColors.secondary;
-            }
-            return null;
-          }),
-        ),
-        radioTheme: RadioThemeData(
-          fillColor: MaterialStateProperty.resolveWith<Color?>((Set<MaterialState> states) {
-            if (states.contains(MaterialState.disabled)) {
-              return null;
-            }
-            if (states.contains(MaterialState.selected)) {
-              return PilllColors.secondary;
-            }
-
-            return null;
-          }),
-        ),
-        checkboxTheme: CheckboxThemeData(
-          fillColor: MaterialStateProperty.resolveWith<Color?>((Set<MaterialState> states) {
-            if (states.contains(MaterialState.disabled)) {
-              return null;
-            }
-            if (states.contains(MaterialState.selected)) {
-              return PilllColors.secondary;
-            }
-            return null;
-          }),
-        ),
-      ),
-      home: ProviderScope(
-        child: RootPage(
-          builder: (_) => ForceUpdate(
-            builder: (_) => UserSignIn(
-              builder: (_, userID) => UserSetup(
-                userID: userID,
-                builder: (_) => InitialSettingOrAppPage(builder: (_, screenType) {
-                  switch (screenType) {
-                    case InitialSettingOrAppPageScreenType.initialSetting:
-                      return ShowPaywallOnAppLaunch(
-                        builder: (_) => SkipInitialSetting(builder: (context, skipInitialSetting) {
-                          if (!skipInitialSetting) {
-                            return InitialSettingPillSheetGroupPageRoute.screen();
-                          } else {
-                            return const HomePage();
-                          }
-                        }),
-                      );
-                    case InitialSettingOrAppPageScreenType.app:
-                      return const HomePage();
-                  }
-                }),
-              ),
-            ),
-          ),
-        ),
-      ),
-      localizationsDelegates: AppLocalizations.localizationsDelegates,
-      supportedLocales: AppLocalizations.supportedLocales,
-      locale: const Locale('ja'),
-    );
-  }
-}
